hyperbahn: refactor all autobahn references to hyperbahn
